+++ task-manager/tests/task.test.js
@@ -1,0 +1,20 @@
+const request = require('supertest')
+const app = require('../src/app')
+const Task = require('../src/models/task')
+const { userOneId, userOne, setupDatabase } = require('./fixtures/db')
+
+// runs before each test case in this test suite and deletes all the users in the testing Database so that you don't run into any errors with duplicates.
+beforeEach(setupDatabase)
+
+test('Should create task for user', async ()=>{
+    const response = await request(app)
+        .post('/tasks')                  // first token, JWT off the object
+        .set('Authorization', `Bearer ${userOne.tokens[0].token}`)
+        .send({
+            description: 'From my test'
+        })
+        .expect(201)
+    const task = await Task.findById(response.body._id)
+    expect(task).not.toBeNull()
+    expect(task.completed).toBe(false)
+})
