+++ task-manager/tests/user.test.js
@@ -1,36 +1,18 @@
 const request = require('supertest')
-const jwt = require('jsonwebtoken')
 const mongoose = require('mongoose')
 const app = require('../src/app')
 const User = require('../src/models/user')
-
-//mongoose will create an id
-const userOneId = new mongoose.Types.ObjectId()
-
-const userOne = {
-    _id: userOneId,
-    name: 'Gingy',
-    email: 'cookieman@example.com',
-    password: '56what!!',
-    tokens: [{
-        token: jwt.sign({ _id: userOneId }, process.env.JWT_SECRET) // creates a new json web token
-    }]
-}
+const { userOneId, userOne, setupDatabase } = require('./fixtures/db')
 
 // Jest setup and teardown
 // Jest lifecycle methods, runs some code just before or after a test case. That will allow you set up some code that clears the database so that your test cases consistently execute as expected.
 
 // runs before each test case in this test suite and deletes all the users in the testing Database so that you don't run into any errors with duplicates.
-beforeEach(async()=>{
-    await User.deleteMany()
-
-    await new User(userOne).save() // creates a user to test with after all the previous ones were deleted.
-})
+beforeEach(setupDatabase)
 
 // runs after each test case in this test suite.
 // afterEach(()=>{
 
-// })
 
 test('Should signup a new user', async()=>{
     const response = await request(app).post('/users').send({
